Update the front end exercise to be more representative of our day-to-day work
With a new round of front-end developer hiring, I've updated the exercise with
some neater visuals, and updated the copy to match.

I've also updated the copy to explain a bit more about what we're looking for in
responses.

I've also updated the job application email address so that it uses the `jobs@`
address, so that the applications go straight into Workable and can be easily
found.