Update icons to latest versions
These icons have since been redrawn and added to the FutureLearn Design System,
so this commit updates the supplied icons to be in sync.

This also fixes an issue where the `icon-create.svg` icon was slightly bigger
than the other icons, meanining it needed its own styling to make it consistent.